bot/requirements.txt
botbuilder-core>=4.14.0
botbuilder-dialogs>=4.15.0
botbuilder-integration-aiohttp>=4.14.0
aiohttp>=3.9
python-dotenv>=1.0
requests